[Refactor] Apply reviews - dividing common constant file to each meaningful files - using html attributes - removing magic number hard coded by making constant
