+++ src/constants/element.js
@@ -3,17 +3,22 @@ export const $ELEMENT = {
   PRICE_INPUT: '#product-price-input',
   QUANTITY_INPUT: '#product-quantity-input',
   ADD_BUTTON: '#product-add-button',
+  PRODUCT_LIST: '.product-inventory-list',
+
   INVENTORY_CONTAINER: '#product-inventory-container',
   APPENDING_CONTAINER: '#appending-product-container',
   CHARGING_CONTAINER: '#charging-money-container',
-  PRODUCT_LIST: '.product-inventory-list',
+
   MENU: '.hash-nav',
   PRODUCT_MANGNE_MENU: '#product-manage-menu',
   VENDING_MACHINE_MANANGE_MENU: '#vending-machine-manage-menu',
   PRODUCT_PURCHASE_MENU: '#product-purchase-menu',
+
+  CHARGE_CONTAINER: '.vending-machine-wrapper',
   CHARGE_INPUT: '#charge-input',
   CHARGE_BUTTON: '#charge-button',
   CHARGE_AMOUNT: '#charge-amount',
+
   CASH_BOX_CONTAINER: '.cashbox-change',
   COIN_500_COUNT: '#coin-500-quantity',
   COIN_100_COUNT: '#coin-100-quantity',
