+++ src/constants/storage.js
@@ -1,0 +1,3 @@
+export const STORAGE = {
+  KEY: 'vending-machine-state',
+};
