+++ src/constants/validate.js
@@ -1,5 +1,6 @@
 export const VALIDATE = {
   MIN_PRICE: 100,
+  MIN_UNIT: 10,
   MIN_NAME_LENGTH: 1,
   MIN_QUANTITY: 1,
 };
