+++ src/js/index.js
@@ -1,4 +1,4 @@
-import { STORAGE } from '../constants/common.js';
+import { STORAGE } from '../constants/storage.js';
 import ChargeController from './controller/chargeController.js';
 import MenuController from './controller/menuController.js';
 import ProductController from './controller/productController.js';
